+++ src/test/java/sw/matilion/foodmart/dao/PositionRepositoryTest.java
@@ -5,6 +5,7 @@ import static org.hamcrest.Matchers.hasSize;
 import static org.hamcrest.Matchers.nullValue;
 import static org.junit.Assert.assertThat;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import org.junit.Test;
@@ -34,15 +35,15 @@ public class PositionRepositoryTest extends AbstractJpaRepositoryTest {
         assertThat(result.get(0).getId(), equalTo(1));
         assertThat(result.get(0).getTitle(), nullValue(String.class));
         assertThat(result.get(0).getPayType(), nullValue(String.class));
-        assertThat(result.get(0).getMinScale(), nullValue(Integer.class));
-        assertThat(result.get(0).getMaxScale(), nullValue(Integer.class));
+        assertThat(result.get(0).getMinScale(), nullValue(BigDecimal.class));
+        assertThat(result.get(0).getMaxScale(), nullValue(BigDecimal.class));
         assertThat(result.get(0).getManagementRole(), nullValue(String.class));
 
         assertThat(result.get(1).getId(), equalTo(2));
         assertThat(result.get(1).getTitle(), equalTo("test_title"));
         assertThat(result.get(1).getPayType(), equalTo("test_type"));
-        assertThat(result.get(1).getMinScale(), equalTo(1));
-        assertThat(result.get(1).getMaxScale(), equalTo(99));
+        assertThat(result.get(1).getMinScale().doubleValue(), equalTo(1d));
+        assertThat(result.get(1).getMaxScale().doubleValue(), equalTo(99d));
         assertThat(result.get(1).getManagementRole(), equalTo("test_role"));
     }
